requirements.txt
Pillow
numpy
spidev
smbus2
bme280
RPi.GPIO